extract eth transaction status logic
